Fixed services to use new config
