+++ src/app.ts
@@ -4,6 +4,7 @@ import { initBackup, restoreBackup } from './services/backup.js'
 import { createServer } from './services/express.js'
 import { loadKeys } from './services/jwt.js'
 import { initSync } from './services/sync.js'
+import { initWebsocket } from './services/websocket.js'
 
 logDebug('Starting server...')
 
@@ -21,4 +22,5 @@ app.listen(port, () => {
 })
 
 initBackup()
+initWebsocket()
 initSync()
